feat: use macro to define static functions
This macro creates `Functions` struct along with a `get` method on it,
it then calls this get method and initializes a static variable once,
after that it also exposes all the functions so that they can be used
directly.

`paste!` is used here to convert snake case of our program to the camel
case that is defined in winapi.

Use `cargo expand icmp::icmp_sys` to see the result of what `bind!`
expands to.